CSS changes + new added files
I have added my image folders, the catalogue.html and cart.html into the projects folder and twiched the header code in all of them so the nav buttons will access them. I have also added the javascript for the catalogue separately from yours @Karolina just in case for now.
I have merged my CSS with
@Karolina's so the style on the catalogue also matched
@Karolina's new header design (which i very much like btw! ) and everything starts to make sense together as a whole - I have though twitched a few things around in the code:
I noticed that when access to a different page (i.e. catalogue) the HOME button on the nav bar stayed black, so it looked like even though you changed pages you are always in the HOME page - i have added some code so now the buttons will automatically get black when changing to the active page, meaning that is the user is on the CATALOGUE page the catalogue button will be black and the HOME button will go grey as the rest.
I have added a certain degree of responsiveness to the design - so you will notice the design will reorganise itself when zooming in and out the page. I am still working though on the responsiveness for mobile ratio hopefully i can manage for the 9th!